requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.7.1
torchmetrics==0.9.3
timm==0.6.7
# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

#--------- torch serve ------#
torchserve==0.6.1
torch-model-archiver==0.6.1
torch-workflow-archiver==0.2.5
ipykernel==6.17.1
captum==0.5.0
torch-tb-profiler==0.4.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# demo

#gradio==3.3.1
pytorch-lightning-bolts==0.3.2.post1

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
